Add PocketService tests; Fix bug
This fixes a bug that previously existed where, if the requested count
was say 5, but one of the 5 results was filtered off because it had a
tag of "private", the resulting list would only have 4 items. With this
change, the result will definitely have the requested amount of items.

I went ahead and fixed it for the pinboard get_bookmarks too since it
was prone to the same problem.

Related to #6.